pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Second trainig project as a backend webdeveloper"
authors = ["Anna Shchankina"]
repository = 'https://github.com/AnnaCanada/python-project-lvl2.git'
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
PyYAML = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:return_diff"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
